% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currency_convert.R
\name{currency_convert}
\alias{currency_convert}
\title{Takes a currency value and the currency type
to be converted to as input and returns the converted currency
value as per the current conversion rate.}
\usage{
currency_convert(value, currency1, currency2)
}
\arguments{
\item{value}{The value of the original currency to be converted}

\item{currency1}{The type of currency originally}

\item{currency2}{The type of currency that the currency1 will be converted to}
}
\value{
Returns converted numeric currency
}
\description{
Takes a currency value and the currency type
to be converted to as input and returns the converted currency
value as per the current conversion rate.
}
\examples{
currency_convert(23, 'USD', 'CAD')
}
