% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastest_slowest_currency.R
\name{fastest_slowest_currency}
\alias{fastest_slowest_currency}
\title{This function takes currency exchange rates data as input and returns a
list of two strings containing the fastest and slowest growing currency
exchange rate in relation to Canadian Dollar.
The data provided contains currency code in the format FX***CAD,
the average exchange rate and the date.}
\usage{
fastest_slowest_currency(start_date, end_date)
}
\arguments{
\item{start_date}{inputted starting date in the format
specified '\%YYYY-\%mm-\%dd'}

\item{end_date}{inputted ending date in the format
specified '\%YYYY-\%mm-\%dd'}
}
\value{
list of lists with fastest and slowest currency and the
current exchange rate of the currencies in the given date range
}
\description{
This function takes currency exchange rates data as input and returns a
list of two strings containing the fastest and slowest growing currency
exchange rate in relation to Canadian Dollar.
The data provided contains currency code in the format FX***CAD,
the average exchange rate and the date.
}
\examples{
fastest_slowest_currency('2019-05-23', '2022-05-30')
}
