% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forxr.R
\name{retrieve_data}
\alias{retrieve_data}
\title{Retrieve historical daily currency exchange rate data for Canadian Dollars in CSV format from the Bank of Canada website.}
\usage{
retrieve_data(start_date, output_path)
}
\arguments{
\item{start_date}{inputted starting date in the format specified '\%YYYY-\%mm-\%dd'}

\item{output_path}{the path where the csv file will get written to}
}
\value{
csv data file that includes historical data on currency exchange rates
}
\description{
Retrieve historical daily currency exchange rate data for Canadian Dollars in CSV format from the Bank of Canada website.
}
\examples{
retrieve_data('2022-05-23', 'my_document/currency_exchange.csv')
}
